+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Added ability to read and write custom invoice notes [115](https://github.com/recurly/recurly-client-php/pull/115)
 * Added vat_location_valid field to Account [127](https://github.com/recurly/recurly-client-php/pull/127)
 * Added updateNotes() and vat_reverse_charge to Subscription. Added vat_reverse_charge_notes to Invoice. [128](https://github.com/recurly/recurly-client-php/pull/128)
+* Added `invoice_number_prefix` and `invoiceNumberWithPrefix()` to `Invoice`. This is to be used with the Country Invoice Sequencing feature. [130](https://github.com/recurly/recurly-client-php/pull/130)
 
 ## Version 2.3.1 (Sept 26th, 2014)